pyproject.toml
[project]
name = "ArtificialInteligence"
version = "0.1.0"
description = "develop integration with ai apis and make some scraping"
authors = [
    {name = "pachecosamuel",email = "pacheco.samuel@hotmail.com"}
]

requires-python = "^3.10"
dependencies = [
    "openai (>=1.61.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "cohere (>=5.13.12,<6.0.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose (>=3.3.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "bcrypt (<4.1)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "firebase (>=4.0.1,<5.0.0)",
    "firebase-admin (>=6.6.0,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
